minor import changes
